refactor: Remove `ReadableTrackingBuffer`. (#647)
`ReadableTrackingBuffer` was no longer used in the `tedious` source
code, only in the unit tests. Instead of keeping the code laying around,
let's get rid of it.
