--- src/tracking-buffer/tracking-buffer.js
@@ -1,2 +1,0 @@
-module.exports.ReadableTrackingBuffer = require('./readable-tracking-buffer');
-module.exports.WritableTrackingBuffer = require('./writable-tracking-buffer');
